[Test] fix mocked functions.
If any function is mocked used stub need to be reverted at some point,
doing something like this

```
before_each(function()
  stub(foo, "validate", function() return false end)
end)

after_each(function()
  foo.validate:revert()
end)

it("Validate stub foo", function()
  assert.same(foo.validate(), false)
end)
```

In APICast all ngx variables are reset on `ngx_helper.lua` but some
methods like `ngx_variables` or `backend_client` are not reset and the
mock function is in there for all other test.

The test in `spec/policy/ngx_variable_spec.lua` validates that
`ngx_variable.available_context` is working correctly, due some other
test can mock this function the test was inestable over the last days.

This commit creates a way to define a cleanup for mocked functions that
are not part of the ngx class.