Tracing: Fix exception with OPENTRACING_FORWARD_HEADER parameter.
Due to the value of a ENV variable in the apicast.conf is a table,
replace can't be used directly on liquid variables.

This change add a new tostring method in the template to be able to use
string filters on nginx templates.

Added a new method because table need to be used to be able to use
empty conditions.

FIX THREESCALE-1660

Signed-off-by: Eloy Coto <eloy.coto@gmail.com>