APICAST: add auth subscribe handlers
At the moment is quite difficult to subscribe to a backend event, so
getting information about apisonator responses in any policy is quite
challenging.

This commit adds two new methods in the request context:
`add_backend_auth_subscriber` and `publish_backend_auth`. When a policy
subscribes to a auth event, each time that publish_backend_auth is
called will execute the callback and policy can retrieve the
information.

Signed-off-by: Eloy Coto <eloy.coto@gmail.com>