Policy: Routing fix URL encoding.
This commit fixes an issue when a special character uses some encoding.
The value to the upstream API if the routing policy was used, is also
encoded, so something like "/foo/test space" should land in APICast as:

`/foo/test%20space/`

And the path to the Upstream API should contain `test%20space`, but
because `ngx.req.set_uri` mark the URL to be encoded again, the URL that
will be received in the Upstream API will be `/test%2520space/`

This commit keeps the URL encoding correctly across all the flow.

Fix THREESCALE-5454

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>