SERVICES_FILTER: Fix issues with oidc.
The oidc-service mapping is not based on a key, so each time that a
service is deleted, the mapping to the given service is broken.

With this change, the OIDC config is stored with the service id, so in
case that some services are removed, it'll keep it correctly mapped.

FIX THREESCALE-6042

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>